+++ src/dependency/core/injection/__init__.py
@@ -1,7 +1,7 @@
 from dependency.core.injection.base import ContainerInjection
 from dependency.core.injection.provider import ProviderInjection
 from dependency.core.injection.injectable import Injectable
-from dependency.core.injection.utils import LazyProvide, LazyProvider, LazyClosing
+from dependency.core.injection.wiring import LazyProvide, LazyProvider, LazyClosing
 
 __all__ = [
     "ContainerInjection",