+++ src/dependency/core/injection/wiring.py
@@ -1,21 +1,6 @@
-from typing import Any, Callable, Generic, Iterable, Optional, TypeVar, Union
+from typing import Any, Callable, Optional, Union
 from dependency_injector import containers, providers
 from dependency_injector.wiring import Modifier, Provide, Provider, Closing
-
-T = TypeVar('T')
-
-class LazyList(Generic[T]):
-    """Lazy List Class for deferred list evaluation.
-    """
-    def __init__(self, iterable: Iterable[T]) -> None:
-        super().__init__()
-        self._iterable = iterable
-        self._list: Optional[list[T]] = None
-
-    def __call__(self, *args: Any, **kwargs: Any) -> list[T]:
-        if self._list is None:
-            self._list = list(self._iterable)
-        return self._list
 
 class LazyWiring:
     """Base Lazy Class for deferred provider resolution.
@@ -29,6 +14,7 @@ class LazyWiring:
 
     @property
     def provider(self) -> Union[providers.Provider[Any], containers.Container, str]:
+        """Return the provider instance."""
         return self._provider()
 
 class LazyProvide(LazyWiring, Provide): # type: ignore